feat: Add motion compensation API

+++ com.rotovr.monitor/com.rotovr.monitor/Common/Model/CompensationModel.cs
@@ -8,12 +8,12 @@ public struct CompensationModel
         Y = 0;
     }
 
-    public CompensationModel(decimal x, decimal y)
+    public CompensationModel(double x, double y)
     {
         X = x;
         Y = y;
     }
 
-    public decimal X { get; }
-    public decimal Y { get; }
+    public double X { get; }
+    public double Y { get; }
 }
